+++ src/state/query/saga.js
@@ -176,8 +176,7 @@ function* getOptions({ target, query }) {
   const options = {}
   if (!_includes(NO_QUERY_LIMIT_TARGETS, target)) {
     Object.assign(options, getTimeFrame(query, target))
-    const getQueryLimit = yield select(getTargetQueryLimit)
-    options.limit = getQueryLimit(target)
+    options.limit = yield select(getTargetQueryLimit, target)
   }
   options.timezone = yield select(getTimezone)
   options.dateFormat = yield select(getDateFormat)
