+++ src/ui/Pagination/Pagination.container.js
@@ -1,13 +1,22 @@
 import { connect } from 'react-redux'
 
 import { getTargetQueryLimit } from 'state/query/selectors'
+import { getPaginationData } from 'state/pagination/selectors'
+import { jumpPage, fetchNext, fetchPrev } from 'state/pagination/actions'
 
 import Pagination from './Pagination'
 
-const mapStateToProps = state => ({
-  getQueryLimit: getTargetQueryLimit(state),
+const mapStateToProps = (state, ownProps) => ({
+  limit: getTargetQueryLimit(state, ownProps.target),
+  ...getPaginationData(state, ownProps.target),
 })
 
-const PaginationContainer = connect(mapStateToProps)(Pagination)
+const mapDispatchToProps = {
+  jumpPage,
+  fetchNext,
+  fetchPrev,
+}
+
+const PaginationContainer = connect(mapStateToProps, mapDispatchToProps)(Pagination)
 
 export default PaginationContainer
