+++ src/ui/Pagination/Pagination.props.js
@@ -1,16 +1,15 @@
 import PropTypes from 'prop-types'
 
 export const propTypes = {
-  dataLen: PropTypes.number.isRequired,
-  getQueryLimit: PropTypes.func.isRequired,
+  target: PropTypes.string.isRequired,
+  entriesSize: PropTypes.number.isRequired,
   loading: PropTypes.bool,
   jumpPage: PropTypes.func,
   offset: PropTypes.number.isRequired,
-  nextClick: PropTypes.func,
-  prevClick: PropTypes.func,
+  fetchNext: PropTypes.func.isRequired,
+  fetchPrev: PropTypes.func.isRequired,
   pageOffset: PropTypes.number.isRequired,
   t: PropTypes.func.isRequired,
-  type: PropTypes.string.isRequired,
   nextPage: PropTypes.oneOfType([
     PropTypes.number,
     PropTypes.bool,
@@ -18,10 +17,6 @@ export const propTypes = {
 }
 
 export const defaultProps = {
-  getQueryLimit: () => {},
-  jumpPage: () => {},
-  nextClick: () => {},
-  prevClick: () => {},
   loading: false,
   nextPage: false,
 }
