--- src/ui/Pagination2/Pagination.container.js
@@ -1,22 +1,0 @@
-import { connect } from 'react-redux'
-
-import { getTargetQueryLimit2 } from 'state/query/selectors'
-import { getPaginationData } from 'state/pagination/selectors'
-import { jumpPage, fetchNext, fetchPrev } from 'state/pagination/actions'
-
-import Pagination from './Pagination'
-
-const mapStateToProps = (state, ownProps) => ({
-  limit: getTargetQueryLimit2(state, ownProps.target),
-  ...getPaginationData(state, ownProps.target),
-})
-
-const mapDispatchToProps = {
-  jumpPage,
-  fetchNext,
-  fetchPrev,
-}
-
-const PaginationContainer = connect(mapStateToProps, mapDispatchToProps)(Pagination)
-
-export default PaginationContainer
