--- src/ui/Pagination2/_Pagination.scss
@@ -1,28 +1,0 @@
-.pagination {
-  &-icon {
-    margin: 4px 2px;
-  }
-
-  &-input {
-    width: 45px;
-    margin-right: 2px;
-    margin-left: 2px;
-    text-align: center;
-  }
-}
-
-@media screen and (max-width: 560px) {
-  .pagination {
-    &-icon {
-      margin: 4px -5px;
-    }
-
-    &-input {
-      width: 35px;
-    }
-
-    .query-limit-selector {
-      margin-left: 5px;
-    }
-  }
-}
