--- src/ui/Pagination2/index.js
@@ -1,3 +1,0 @@
-import PaginationContainer from './Pagination.container'
-
-export default PaginationContainer
