+++ src/ui/QueryLimitSelector/QueryLimitSelector.props.js
@@ -1,14 +1,13 @@
 import PropTypes from 'prop-types'
 
 export const propTypes = {
-  getQueryLimit: PropTypes.func.isRequired,
+  queryLimit: PropTypes.number.isRequired,
   setQueryLimit: PropTypes.func.isRequired,
   t: PropTypes.func.isRequired,
   target: PropTypes.string.isRequired,
 }
 
 export const defaultProps = {
-  getQueryLimit: () => {},
   setQueryLimit: () => {},
   target: '',
 }
